backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
openai==2.21.0
python-dotenv==1.2.1
supabase==2.28.0
numpy==2.4.2
jinja2==3.1.6
playwright==1.58.0
